Remove post-destroy component, refactor first-task component
The post-destroy component was removed as it was deemed unnecessary. Transferred the functionality to destroy posts under first-task. This refactor simplifies the application structure and reduces the amount of code, making it cleaner. Also, brought minor changes in HTML to replace custom CSS being applied to checkbox and used Angular Material's checkbox component. This brings uniformity across the application and leverages Angular Material's default styling.